+++ src/main/java/saigonwithlove/ivy/intellij/shared/Configuration.java
@@ -1,13 +1,22 @@
 package saigonwithlove.ivy.intellij.shared;
 
-import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
+import org.apache.commons.lang3.StringUtils;
 
 @Getter
-@Setter
-@AllArgsConstructor
+@Builder
 public class Configuration {
   private String name;
-  private String value;
+  @Setter private String value;
+  private String defaultValue;
+
+  public boolean isModified() {
+    return value != null && !StringUtils.equals(value, defaultValue);
+  }
+
+  public String getValue() {
+    return value != null ? value : defaultValue;
+  }
 }