+++ src/main/java/saigonwithlove/ivy/intellij/shared/Configurations.java
@@ -3,10 +3,13 @@ package saigonwithlove.ivy.intellij.shared;
 import com.google.common.base.Preconditions;
 import com.intellij.openapi.vfs.VirtualFile;
 import java.io.IOException;
+import java.util.List;
+import java.util.Map;
 import java.util.NoSuchElementException;
 import java.util.Optional;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+import java.util.stream.Collectors;
 import lombok.experimental.UtilityClass;
 import org.apache.commons.lang3.StringUtils;
 import org.jetbrains.annotations.NotNull;
@@ -15,7 +18,8 @@ import org.jetbrains.annotations.NotNull;
 public class Configurations {
   private static final Pattern DEFAULT_VALUE_FILE_PATTERN =
       Pattern.compile("^val.*guid=\\{([0-9A-Z]+)\\}.*isdefault=\\{true\\}", Pattern.MULTILINE);
-  private static final Pattern VALUE_PATTERN = Pattern.compile("/value \"(.*)\"/description", Pattern.MULTILINE);
+  private static final Pattern VALUE_PATTERN =
+      Pattern.compile("/value \"(.*)\"/description", Pattern.MULTILINE);
 
   @NotNull
   public static Configuration parseGlobalVariable(@NotNull VirtualFile globalVariableDirectory) {
@@ -25,7 +29,7 @@ public class Configurations {
             .map(Configurations::getValueFile)
             .map(Configurations::getValue)
             .orElse(StringUtils.EMPTY);
-    return new Configuration(name, value);
+    return Configuration.builder().name(name).defaultValue(value).build();
   }
 
   @NotNull
@@ -61,5 +65,20 @@ public class Configurations {
       throw new RuntimeException("Could not read value of global variable: " + valueFile);
     }
     return StringUtils.EMPTY;
+  }
+
+  @NotNull
+  public List<Configuration> buildConfigurations(
+      Map<String, String> defaultConfigurations, Map<String, String> modifiedConfigurations) {
+    return defaultConfigurations.entrySet().stream()
+        .map(
+            entry ->
+                Configuration.builder()
+                    .name(entry.getKey())
+                    .defaultValue(entry.getValue())
+                    .value(modifiedConfigurations.get(entry.getKey()))
+                    .build())
+        .sorted((a, b) -> a.getName().compareToIgnoreCase(b.getName()))
+        .collect(Collectors.toList());
   }
 }